interface as sub module
