+++ audbackend/__init__.py
@@ -1,3 +1,4 @@
+from audbackend import interface
 from audbackend.core.api import access
 from audbackend.core.api import available
 from audbackend.core.api import create
@@ -6,9 +7,6 @@ from audbackend.core.api import register
 from audbackend.core.backend import Backend
 from audbackend.core.errors import BackendError
 from audbackend.core.filesystem import FileSystem
-from audbackend.core.interface import Interface
-from audbackend.core.interface import Unversioned
-from audbackend.core.interface import Versioned
 from audbackend.core.repository import Repository
 
 # Import optional backends
