+++ docs/api-src/audbackend.rst
@@ -27,29 +27,6 @@ backend by deriving from
 
     Backend
 
-To access a backend users
-can choose between two interfaces
-:class:`audbackend.Unversioned`
-or
-:class:`audbackend.Versioned`:
-
-.. autosummary::
-    :toctree:
-    :nosignatures:
-
-    Unversioned
-    Versioned
-
-Users can implement their own
-interface by deriving from
-:class:`audbackend.Interface`
-
-.. autosummary::
-    :toctree:
-    :nosignatures:
-
-    Interface
-
 In addition to the backends
 the following classes and functions are available.
